airflow-docker/requirements.txt
pandas
numpy
pymongo
openaq
requests
joblib
scikit-learn
python-dateutil
seaborn